Just move the PSD export to Compiler from FormatZIP
