+++ FormEdit.cs
@@ -5019,7 +5019,7 @@ namespace IfcDoc
 						{
 							foreach (DocPropertySet docPset in docSchema.PropertySets)
 							{
-								PropertySetDef psd = FormatZIP.ExportPsd(docPset, mapPropEnum, this.m_project);
+								PropertySetDef psd = Compiler.ExportPsd(docPset, mapPropEnum, this.m_project);
 								string filename = System.IO.Path.Combine(this.folderBrowserDialog.SelectedPath, docPset.Name + ".xml");
 								using (FormatXML format = new FormatXML(filename, typeof(PropertySetDef)))//, PropertySetDef.DefaultNamespace))
 								{
@@ -5030,7 +5030,7 @@ namespace IfcDoc
 
 							foreach (DocQuantitySet docQset in docSchema.QuantitySets)
 							{
-								QtoSetDef qto = FormatZIP.ExportQto(docQset, this.m_project);
+								QtoSetDef qto = Compiler.ExportQto(docQset, this.m_project);
 								string filename = System.IO.Path.Combine(this.folderBrowserDialog.SelectedPath, docQset.Name + ".xml");
 								using (FormatXML format = new FormatXML(filename, typeof(QtoSetDef), QtoSetDef.DefaultNamespace))
 								{
